Refactor user and usergame components folder structure
