+++ components/usergame/usergame-delete-dialog.tsx
